Activity 6
Created a js script to receive a number from 1-10 from the user and return the name of a shape that has that many sides and linked a html page to it called name_the_polygon.html. Also moved brand_humble_jaeger.html to the brand folder and will be moving first_scripts.html in there as well after receiving a grade for it.